+++ lib/commanded/aggregates/multi.ex
@@ -46,58 +46,86 @@ defmodule Commanded.Aggregate.Multi do
         end
         defp check_balance(%BankAccount{}), do: []
       end
+
   """
 
   alias Commanded.Aggregate.Multi
 
   @type t :: %__MODULE__{
-    aggregate: struct(),
-    executions: list(function()),
-  }
+          aggregate: struct(),
+          executions: list(function())
+        }
 
-  defstruct [
-    aggregate: nil,
-    executions: [],
-  ]
+  defstruct [:aggregate, executions: []]
 
   @doc """
   Create a new `Commanded.Aggregate.Multi` struct.
   """
-  @spec new(aggregate :: struct()) :: Multi.t
+  @spec new(aggregate :: struct()) :: Multi.t()
   def new(aggregate), do: %Multi{aggregate: aggregate}
 
   @doc """
   Adds a command execute function to the multi.
   """
-  @spec execute(Multi.t, function()) :: Multi.t
-  def execute(%Multi{executions: executions} = multi, execute_fun)
-    when is_function(execute_fun)
-  do
-    %Multi{multi |
-      executions: [execute_fun | executions],
-    }
+  @spec execute(Multi.t(), function()) :: Multi.t()
+  def execute(%Multi{} = multi, execute_fun) when is_function(execute_fun, 1) do
+    %Multi{executions: executions} = multi
+
+    %Multi{multi | executions: [execute_fun | executions]}
+  end
+
+  @doc """
+  Reduce an enumerable by executing the function for each item.
+
+  The aggregate `apply/2` function will be called after each event returned by
+  the execute function. This allows you to calculate values from the aggregate
+  state based upon events produced by previous items in the enumerable, such as
+  running totals.
+
+  ## Example
+
+      alias Commanded.Aggregate.Multi
+
+      aggregate
+      |> Multi.new()
+      |> Multi.reduce([1, 2, 3], fn aggregate, item ->
+        %AnEvent{item: item, total: aggregate.total + item}
+      end)
+
+  """
+  @spec reduce(Multi.t(), Enum.t(), function()) :: Multi.t()
+  def reduce(%Multi{} = multi, enumerable, execute_fun) when is_function(execute_fun, 2) do
+    Enum.reduce(enumerable, multi, fn item, %Multi{} = multi ->
+      execute(multi, &execute_fun.(&1, item))
+    end)
   end
 
   @doc """
   Run the execute functions contained within the multi, returning the updated
-  aggregate state and any created events.
+  aggregate state and all created events.
   """
-  @spec run(Multi.t) :: {aggregate :: struct(), list(event :: struct())} | {:error, reason :: any()}
+  @spec run(Multi.t()) ::
+          {aggregate :: struct(), list(event :: struct())} | {:error, reason :: any()}
   def run(%Multi{aggregate: aggregate, executions: executions}) do
     try do
       executions
       |> Enum.reverse()
-      |> Enum.reduce({aggregate, []}, fn (execute_fun, {aggregate, events}) ->
-        pending_events =
-          case execute_fun.(aggregate) do
-            {:error, _reason} = error -> throw(error)
-            pending_events -> List.wrap(pending_events)
-          end
+      |> Enum.reduce({aggregate, []}, fn execute_fun, {aggregate, events} ->
+        case execute_fun.(aggregate) do
+          {:error, _reason} = error ->
+            throw(error)
 
-        {apply_events(aggregate, pending_events), events ++ pending_events}
+          %Multi{} = multi ->
+            Multi.run(multi)
+
+          pending_events ->
+            pending_events = List.wrap(pending_events)
+
+            {apply_events(aggregate, pending_events), events ++ pending_events}
+        end
       end)
     catch
-      {:error, _reason} = error -> error
+      {:error, _error} = error -> error
     end
   end